Serverless/annotateAPC/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.22.0
google-cloud-error-reporting==0.34.0
google-cloud-bigquery==1.22.0
